Move substream modifiers next to other substream code.
