+++ src/card-controller/query-string-manager.ts
@@ -1,3 +1,4 @@
+import { SubstreamSelectViewModifier } from '../components-lib/live/substream';
 import { ViewActionConfig } from '../config/schema/actions/custom/view';
 import { AdvancedCameraCardCustomActionConfig } from '../config/schema/actions/types';
 import {
@@ -6,7 +7,6 @@ import {
   createViewAction,
 } from '../utils/action.js';
 import { CardQueryStringAPI } from './types';
-import { SubstreamSelectViewModifier } from './view/modifiers/substream-select';
 import { ViewParametersUserSpecified } from './view/types.js';
 
 interface QueryStringViewIntent {
