--- src/card-controller/view/modifiers/substream-off.ts
@@ -1,9 +1,0 @@
-import { removeSubstream } from '../../../utils/substream';
-import { View } from '../../../view/view';
-import { ViewModifier } from '../types';
-
-export class SubstreamOffViewModifier implements ViewModifier {
-  public modify(view: View): void {
-    removeSubstream(view);
-  }
-}
