+++ src/card-controller/view/view-manager.ts
@@ -1,8 +1,8 @@
 import { ViewContext } from 'view';
+import { getStreamCameraID } from '../../components-lib/live/substream';
 import { log } from '../../utils/debug';
-import { getStreamCameraID } from '../../utils/substream';
-import { View } from '../../view/view';
 import { getViewTargetID } from '../../view/target-id';
+import { View } from '../../view/view';
 import { InitializationAspect } from '../initialization-manager';
 import { CardViewAPI } from '../types';
 import { ViewFactory } from './factory';
