+++ tests/card-controller/actions/actions/substream-on.test.ts
@@ -1,7 +1,7 @@
 import { expect, it } from 'vitest';
 import { SubstreamOnAction } from '../../../../src/card-controller/actions/actions/substream-on';
+import { SubstreamOnViewModifier } from '../../../../src/components-lib/live/substream';
 import { createCardAPI } from '../../../test-utils';
-import { SubstreamOnViewModifier } from '../../../../src/card-controller/view/modifiers/substream-on';
 
 it('should handle live_substream_on action', async () => {
   const api = createCardAPI();
