+++ tests/card-controller/actions/actions/substream-select.test.ts
@@ -1,7 +1,7 @@
 import { expect, it } from 'vitest';
 import { SubstreamSelectAction } from '../../../../src/card-controller/actions/actions/substream-select';
+import { SubstreamSelectViewModifier } from '../../../../src/components-lib/live/substream';
 import { createCardAPI } from '../../../test-utils';
-import { SubstreamSelectViewModifier } from '../../../../src/card-controller/view/modifiers/substream-select';
 
 it('should handle live_substream_select action', async () => {
   const api = createCardAPI();
@@ -16,9 +16,7 @@ it('should handle live_substream_select action', async () => {
 
   await action.execute(api);
 
-  expect(api.getViewManager().setViewByParameters).toBeCalledWith(
-    expect.objectContaining({
-      modifiers: expect.arrayContaining([expect.any(SubstreamSelectViewModifier)]),
-    }),
-  );
+  expect(api.getViewManager().setViewByParameters).toBeCalledWith({
+    modifiers: [expect.any(SubstreamSelectViewModifier)],
+  });
 });
