+++ tests/card-controller/query-string-manager.test.ts
@@ -2,7 +2,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 import { CardController } from '../../src/card-controller/controller';
 import { QueryStringManager } from '../../src/card-controller/query-string-manager';
-import { SubstreamSelectViewModifier } from '../../src/card-controller/view/modifiers/substream-select';
+import { SubstreamSelectViewModifier } from '../../src/components-lib/live/substream';
 import { createCardAPI, createConfig } from '../test-utils';
 
 const setQueryString = (qs: string): void => {
